+++ src/components/layout/Sidebar/Sidebar.tsx
@@ -1,7 +1,7 @@
 import { mergeClassNames } from '@/utils/mergeClassNames';
 import PrimarySidebar from './SidebarNav/SidebarNav';
 import SecondarySidebar from './SidebarPanel/SidebarPanel';
-import { useSidebarStore } from '@/stores/usesidebarStore';
+import { useSidebarStore } from '@/stores/useSidebarStore';
 import StellarDataBinder from '@/components/stellarDataBinder/StellarDataBinder';
 
 interface SidebarProps {
