+++ src/components/layout/Sidebar/SidebarPanel/Profile/LikesPanel.tsx
@@ -1,4 +1,4 @@
-import { useProfileStore } from '@/stores/useprofileStore';
+import { useProfileStore } from '@/stores/useProfileStore';
 import PanelHeader from '../PanelHeader';
 import CardItem from '@/components/panel/galaxy/community/CardItem';
 import { ScrollArea } from '@/components/common/Scrollarea';
