+++ src/components/panel/galaxy/community/List.tsx
@@ -12,7 +12,7 @@ import { ErrorBoundary } from 'react-error-boundary';
 import LoadingIcon from '@/components/common/LoadingIcon';
 import { useSelectedStellarStore } from '@/stores/useSelectedStellarStore';
 import { useStellarTabStore } from '@/stores/useStellarTabStore';
-import { useSidebarStore } from '@/stores/usesidebarStore';
+import { useSidebarStore } from '@/stores/useSidebarStore';
 import useStellarSystemSelection from '@/hooks/useStellarSystemSelection';
 
 const GALAXY_LIST_LIMIT = 3;
