+++ src/hooks/api/useStellar.ts
@@ -6,7 +6,7 @@ import type { StellarType } from '@/types/stellar';
 import { useStellarStore } from '@/stores/useStellarStore';
 import { useEffect } from 'react';
 import { useSelectedStellarStore } from '@/stores/useSelectedStellarStore';
-import { useSidebarStore } from '@/stores/sidebarStore';
+import { useSidebarStore } from '@/stores/useSidebarStore';
 import { useStellarTabStore } from '@/stores/useStellarTabStore';
 
 // 생성
